% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodal_visualization.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize first letter of each word}
\usage{
capitalize(x)
}
\arguments{
\item{x}{Character string to capitalize}
}
\value{
Capitalized string
}
\description{
Capitalize first letter of each word
}
\keyword{internal}
