% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{CharacterVoiceFactory}
\alias{CharacterVoiceFactory}
\title{Character Voice Factory}
\description{
Factory class to create character voice instances
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{voices}}{List of available voice instances}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CharacterVoiceFactory-new}{\code{CharacterVoiceFactory$new()}}
\item \href{#method-CharacterVoiceFactory-get_voice}{\code{CharacterVoiceFactory$get_voice()}}
\item \href{#method-CharacterVoiceFactory-list_voices}{\code{CharacterVoiceFactory$list_voices()}}
\item \href{#method-CharacterVoiceFactory-clone}{\code{CharacterVoiceFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacterVoiceFactory-new"></a>}}
\if{latex}{\out{\hypertarget{method-CharacterVoiceFactory-new}{}}}
\subsection{Method \code{new()}}{
Initialize the factory with available voices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterVoiceFactory$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacterVoiceFactory-get_voice"></a>}}
\if{latex}{\out{\hypertarget{method-CharacterVoiceFactory-get_voice}{}}}
\subsection{Method \code{get_voice()}}{
Get a voice instance by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterVoiceFactory$get_voice(voice_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{voice_name}}{Name of the voice to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character voice instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacterVoiceFactory-list_voices"></a>}}
\if{latex}{\out{\hypertarget{method-CharacterVoiceFactory-list_voices}{}}}
\subsection{Method \code{list_voices()}}{
List all available voices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterVoiceFactory$list_voices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with voice names and descriptions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacterVoiceFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CharacterVoiceFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterVoiceFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
