% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{ContextAnalyzer}
\alias{ContextAnalyzer}
\title{Context Analyzer Class}
\description{
R6 Class for analyzing R environment context
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{context_data}}{A list containing analyzed context data}

\item{\code{privacy_settings}}{A list containing privacy configuration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContextAnalyzer-new}{\code{ContextAnalyzer$new()}}
\item \href{#method-ContextAnalyzer-analyze_environment}{\code{ContextAnalyzer$analyze_environment()}}
\item \href{#method-ContextAnalyzer-analyze_data_frames}{\code{ContextAnalyzer$analyze_data_frames()}}
\item \href{#method-ContextAnalyzer-analyze_packages}{\code{ContextAnalyzer$analyze_packages()}}
\item \href{#method-ContextAnalyzer-analyze_command_history}{\code{ContextAnalyzer$analyze_command_history()}}
\item \href{#method-ContextAnalyzer-analyze_environment_info}{\code{ContextAnalyzer$analyze_environment_info()}}
\item \href{#method-ContextAnalyzer-score_data_frame_relevance}{\code{ContextAnalyzer$score_data_frame_relevance()}}
\item \href{#method-ContextAnalyzer-extract_objects_from_history}{\code{ContextAnalyzer$extract_objects_from_history()}}
\item \href{#method-ContextAnalyzer-generate_contextual_examples}{\code{ContextAnalyzer$generate_contextual_examples()}}
\item \href{#method-ContextAnalyzer-generate_dplyr_example}{\code{ContextAnalyzer$generate_dplyr_example()}}
\item \href{#method-ContextAnalyzer-generate_ggplot_example}{\code{ContextAnalyzer$generate_ggplot_example()}}
\item \href{#method-ContextAnalyzer-generate_stats_example}{\code{ContextAnalyzer$generate_stats_example()}}
\item \href{#method-ContextAnalyzer-generate_tidyr_example}{\code{ContextAnalyzer$generate_tidyr_example()}}
\item \href{#method-ContextAnalyzer-generate_purrr_example}{\code{ContextAnalyzer$generate_purrr_example()}}
\item \href{#method-ContextAnalyzer-generate_survival_example}{\code{ContextAnalyzer$generate_survival_example()}}
\item \href{#method-ContextAnalyzer-generate_stats_model_example}{\code{ContextAnalyzer$generate_stats_model_example()}}
\item \href{#method-ContextAnalyzer-generate_data_table_example}{\code{ContextAnalyzer$generate_data_table_example()}}
\item \href{#method-ContextAnalyzer-generate_ml_example}{\code{ContextAnalyzer$generate_ml_example()}}
\item \href{#method-ContextAnalyzer-format_context_for_prompt}{\code{ContextAnalyzer$format_context_for_prompt()}}
\item \href{#method-ContextAnalyzer-suggest_next_steps}{\code{ContextAnalyzer$suggest_next_steps()}}
\item \href{#method-ContextAnalyzer-generate_stringr_example}{\code{ContextAnalyzer$generate_stringr_example()}}
\item \href{#method-ContextAnalyzer-generate_lubridate_example}{\code{ContextAnalyzer$generate_lubridate_example()}}
\item \href{#method-ContextAnalyzer-clone}{\code{ContextAnalyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ContextAnalyzer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$new(privacy_settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{privacy_settings}}{List of privacy settings}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-analyze_environment"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-analyze_environment}{}}}
\subsection{Method \code{analyze_environment()}}{
Analyze the current R environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$analyze_environment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self (for method chaining)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-analyze_data_frames"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-analyze_data_frames}{}}}
\subsection{Method \code{analyze_data_frames()}}{
Analyze data frames in the global environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$analyze_data_frames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self (for method chaining)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-analyze_packages"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-analyze_packages}{}}}
\subsection{Method \code{analyze_packages()}}{
Analyze packages and recently used functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$analyze_packages()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self (for method chaining)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-analyze_command_history"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-analyze_command_history}{}}}
\subsection{Method \code{analyze_command_history()}}{
Analyze command history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$analyze_command_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self (for method chaining)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-analyze_environment_info"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-analyze_environment_info}{}}}
\subsection{Method \code{analyze_environment_info()}}{
Analyze general environment information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$analyze_environment_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self (for method chaining)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-score_data_frame_relevance"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-score_data_frame_relevance}{}}}
\subsection{Method \code{score_data_frame_relevance()}}{
Score data frames for relevance to a given function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$score_data_frame_relevance(func_name, func_metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{func_metadata}}{Metadata about the function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named vector of relevance scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-extract_objects_from_history"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-extract_objects_from_history}{}}}
\subsection{Method \code{extract_objects_from_history()}}{
Extract objects used in recent commands
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$extract_objects_from_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of object names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_contextual_examples"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_contextual_examples}{}}}
\subsection{Method \code{generate_contextual_examples()}}{
Generate function examples using actual data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_contextual_examples(
  func_name,
  func_metadata,
  top_n = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{func_metadata}}{Metadata about the function}

\item{\code{top_n}}{Number of top data frames to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of example strings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_dplyr_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_dplyr_example}{}}}
\subsection{Method \code{generate_dplyr_example()}}{
Generate example for dplyr functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_dplyr_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_ggplot_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_ggplot_example}{}}}
\subsection{Method \code{generate_ggplot_example()}}{
Generate example for ggplot2 functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_ggplot_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_stats_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_stats_example}{}}}
\subsection{Method \code{generate_stats_example()}}{
Generate example for basic statistics functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_stats_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_tidyr_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_tidyr_example}{}}}
\subsection{Method \code{generate_tidyr_example()}}{
Generate example for statistical modeling functions


Generate example for tidyr functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_tidyr_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_purrr_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_purrr_example}{}}}
\subsection{Method \code{generate_purrr_example()}}{
Generate example for purrr functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_purrr_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_survival_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_survival_example}{}}}
\subsection{Method \code{generate_survival_example()}}{
Generate example for survival analysis functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_survival_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_stats_model_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_stats_model_example}{}}}
\subsection{Method \code{generate_stats_model_example()}}{
Generate example for statistical modeling functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_stats_model_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_data_table_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_data_table_example}{}}}
\subsection{Method \code{generate_data_table_example()}}{
Generate example for data.table functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_data_table_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_ml_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_ml_example}{}}}
\subsection{Method \code{generate_ml_example()}}{
Generate example for machine learning functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_ml_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-format_context_for_prompt"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-format_context_for_prompt}{}}}
\subsection{Method \code{format_context_for_prompt()}}{
Format context data for inclusion in AI prompt
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$format_context_for_prompt(func_name, func_metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{func_metadata}}{Metadata about the function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Formatted context string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-suggest_next_steps"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-suggest_next_steps}{}}}
\subsection{Method \code{suggest_next_steps()}}{
Suggest next logical steps based on context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$suggest_next_steps(func_name, func_metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{func_metadata}}{Metadata about the function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of suggested next steps
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_stringr_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_stringr_example}{}}}
\subsection{Method \code{generate_stringr_example()}}{
Generate example for stringr text manipulation functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_stringr_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-generate_lubridate_example"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-generate_lubridate_example}{}}}
\subsection{Method \code{generate_lubridate_example()}}{
Generate example for lubridate date/time functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$generate_lubridate_example(func_name, df_name, df_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func_name}}{Name of the function}

\item{\code{df_name}}{Name of the data frame}

\item{\code{df_info}}{Information about the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Example string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
