% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{CynicalDetectiveVoice}
\alias{CynicalDetectiveVoice}
\title{Cynical Detective Character Voice}
\description{
A character voice that's skeptical, analytical, and straight-talking
}
\keyword{internal}
\section{Super class}{
\code{\link[tldrAI:CharacterVoice]{tldrAI::CharacterVoice}} -> \code{CynicalDetectiveVoice}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CynicalDetectiveVoice-new}{\code{CynicalDetectiveVoice$new()}}
\item \href{#method-CynicalDetectiveVoice-transform}{\code{CynicalDetectiveVoice$transform()}}
\item \href{#method-CynicalDetectiveVoice-transform_purpose}{\code{CynicalDetectiveVoice$transform_purpose()}}
\item \href{#method-CynicalDetectiveVoice-clone}{\code{CynicalDetectiveVoice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CynicalDetectiveVoice-new"></a>}}
\if{latex}{\out{\hypertarget{method-CynicalDetectiveVoice-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Cynical Detective voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CynicalDetectiveVoice$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CynicalDetectiveVoice-transform"></a>}}
\if{latex}{\out{\hypertarget{method-CynicalDetectiveVoice-transform}{}}}
\subsection{Method \code{transform()}}{
Transform a response using Cynical Detective voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CynicalDetectiveVoice$transform(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The original AI response text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CynicalDetectiveVoice-transform_purpose"></a>}}
\if{latex}{\out{\hypertarget{method-CynicalDetectiveVoice-transform_purpose}{}}}
\subsection{Method \code{transform_purpose()}}{
Transform just the purpose content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CynicalDetectiveVoice$transform_purpose(purpose_lines)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{purpose_lines}}{Character vector with purpose content lines}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed purpose lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CynicalDetectiveVoice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CynicalDetectiveVoice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CynicalDetectiveVoice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
