% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{EnthusiasticExplorerVoice}
\alias{EnthusiasticExplorerVoice}
\title{Enthusiastic Explorer Character Voice}
\description{
A character voice with enthusiasm, excitement, and curiosity
}
\keyword{internal}
\section{Super class}{
\code{\link[tldrAI:CharacterVoice]{tldrAI::CharacterVoice}} -> \code{EnthusiasticExplorerVoice}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EnthusiasticExplorerVoice-new}{\code{EnthusiasticExplorerVoice$new()}}
\item \href{#method-EnthusiasticExplorerVoice-transform}{\code{EnthusiasticExplorerVoice$transform()}}
\item \href{#method-EnthusiasticExplorerVoice-transform_purpose}{\code{EnthusiasticExplorerVoice$transform_purpose()}}
\item \href{#method-EnthusiasticExplorerVoice-clone}{\code{EnthusiasticExplorerVoice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnthusiasticExplorerVoice-new"></a>}}
\if{latex}{\out{\hypertarget{method-EnthusiasticExplorerVoice-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Enthusiastic Explorer voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnthusiasticExplorerVoice$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnthusiasticExplorerVoice-transform"></a>}}
\if{latex}{\out{\hypertarget{method-EnthusiasticExplorerVoice-transform}{}}}
\subsection{Method \code{transform()}}{
Transform a response using Enthusiastic Explorer voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnthusiasticExplorerVoice$transform(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The original AI response text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnthusiasticExplorerVoice-transform_purpose"></a>}}
\if{latex}{\out{\hypertarget{method-EnthusiasticExplorerVoice-transform_purpose}{}}}
\subsection{Method \code{transform_purpose()}}{
Transform just the purpose content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnthusiasticExplorerVoice$transform_purpose(purpose_lines)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{purpose_lines}}{Character vector with purpose content lines}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed purpose lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnthusiasticExplorerVoice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EnthusiasticExplorerVoice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnthusiasticExplorerVoice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
