% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{extract_code_block}
\alias{extract_code_block}
\title{Extract the content from code blocks in the response}
\usage{
extract_code_block(response)
}
\arguments{
\item{response}{The raw response from the AI}
}
\value{
Character string with the content from the code block
}
\description{
Extract the content from code blocks in the response
}
\keyword{internal}
