% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{extract_functions_from_code}
\alias{extract_functions_from_code}
\title{Extract function calls from R code}
\usage{
extract_functions_from_code(code)
}
\arguments{
\item{code}{The R code to analyze}
}
\value{
Character vector of function names
}
\description{
Extract function calls from R code
}
\keyword{internal}
