% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{extract_variables_from_code}
\alias{extract_variables_from_code}
\title{Extract variables from R code}
\usage{
extract_variables_from_code(code)
}
\arguments{
\item{code}{The R code to analyze}
}
\value{
Character vector of variable names
}
\description{
Extract variables from R code
}
\keyword{internal}
