% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R, R/metadata.R
\name{find_package}
\alias{find_package}
\title{Find the package that a function belongs to}
\usage{
find_package(func_name)

find_package(func_name)
}
\arguments{
\item{func_name}{The name of the function}
}
\value{
Character string with the package name

Character string with the package name
}
\description{
Find the package that a function belongs to

Find the package that a function belongs to
}
\keyword{internal}
