% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format_content}
\alias{format_content}
\title{Format the content for display}
\usage{
format_content(content)
}
\arguments{
\item{content}{The content to format}
}
\value{
Character string with the formatted content
}
\description{
Format the content for display
}
\keyword{internal}
