% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Get a configuration value}
\usage{
get_config(key, default = NULL)
}
\arguments{
\item{key}{The configuration key to retrieve}

\item{default}{The default value if the key is not found}
}
\value{
The configuration value
}
\description{
Get a configuration value
}
\keyword{internal}
