% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_all}
\alias{get_config_all}
\title{Get all configuration values}
\usage{
get_config_all()
}
\value{
A list of all configuration values
}
\description{
Get all configuration values
}
\keyword{internal}
