% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_path}
\alias{get_config_path}
\title{Get the path to the configuration file}
\usage{
get_config_path()
}
\value{
Character string with the path to the configuration file
}
\description{
Get the path to the configuration file
}
\keyword{internal}
