% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodal_visualization.R
\name{get_example_code}
\alias{get_example_code}
\title{Get example code for the function}
\usage{
get_example_code(func_name, is_dplyr, is_filter)
}
\arguments{
\item{func_name}{Function name}

\item{is_dplyr}{Is it a dplyr function?}

\item{is_filter}{Is it a filter-like function?}
}
\value{
Character string with example code
}
\description{
Get example code for the function
}
\keyword{internal}
