% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_function_description}
\alias{get_function_description}
\title{Get the function description from help pages}
\usage{
get_function_description(func_name, package)
}
\arguments{
\item{func_name}{The name of the function}

\item{package}{The package name}
}
\value{
Character string with the function description
}
\description{
Get the function description from help pages
}
\keyword{internal}
