% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_function_metadata}
\alias{get_function_metadata}
\title{Get metadata about an R function}
\usage{
get_function_metadata(func_name, package = NULL)
}
\arguments{
\item{func_name}{The name of the function}

\item{package}{The package name (optional)}
}
\value{
A list with function metadata, or a special metadata object for missing packages
}
\description{
Get metadata about an R function
}
\keyword{internal}
