% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_function_signature}
\alias{get_function_signature}
\title{Get the function signature as a string}
\usage{
get_function_signature(func_name, func)
}
\arguments{
\item{func_name}{The name of the function}

\item{func}{The function object}
}
\value{
Character string with the formatted signature
}
\description{
Get the function signature as a string
}
\keyword{internal}
