% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodal_visualization.R
\name{get_function_source}
\alias{get_function_source}
\title{Get function source code}
\usage{
get_function_source(func_name, pkg, args_str)
}
\arguments{
\item{func_name}{Function name}

\item{pkg}{Package name}

\item{args_str}{Arguments string for placeholder}
}
\value{
Character string of function source code
}
\description{
Get function source code
}
\keyword{internal}
