% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{get_missing_package_template}
\alias{get_missing_package_template}
\title{Get template for missing package}
\usage{
get_missing_package_template()
}
\value{
Character string containing the prompt template for missing packages
}
\description{
Get template for missing package
}
\keyword{internal}
