% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/context.R, R/explain.R,
%   R/multimodal_visualization.R, R/output.R, R/tldr.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y

x \%||\% y

x \%||\% y

x \%||\% y

x \%||\% y

x \%||\% y
}
\description{
Null-coalescing operator

Null-coalescing operator

Null-coalescing operator

Null coalescing operator

Null-coalescing operator

Null-coalescing operator
}
\keyword{internal}
