% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{LLMProviderFactory}
\alias{LLMProviderFactory}
\title{LLM Provider Factory}
\description{
Factory class to create appropriate LLM providers
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LLMProviderFactory-create_provider}{\code{LLMProviderFactory$create_provider()}}
\item \href{#method-LLMProviderFactory-clone}{\code{LLMProviderFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LLMProviderFactory-create_provider"></a>}}
\if{latex}{\out{\hypertarget{method-LLMProviderFactory-create_provider}{}}}
\subsection{Method \code{create_provider()}}{
Create a provider instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LLMProviderFactory$create_provider(provider_name, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider_name}}{Name of the provider}

\item{\code{config}}{Configuration}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Provider instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LLMProviderFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LLMProviderFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LLMProviderFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
