% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{print_visualization}
\alias{print_visualization}
\title{Print a visualization}
\usage{
print_visualization(visualization)
}
\arguments{
\item{visualization}{A visualization handler object}
}
\value{
Invisibly returns the visualization handler
}
\description{
Print a visualization
}
\examples{
\dontrun{
metadata <- get_function_metadata("mean")
vis <- create_visualization("mean", metadata)
print_visualization(vis)
}
}
