% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{ReluctantHelperVoice}
\alias{ReluctantHelperVoice}
\title{Reluctant Helper Character Voice}
\description{
A character voice that acts comically put-upon about providing help
}
\keyword{internal}
\section{Super class}{
\code{\link[tldrAI:CharacterVoice]{tldrAI::CharacterVoice}} -> \code{ReluctantHelperVoice}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReluctantHelperVoice-new}{\code{ReluctantHelperVoice$new()}}
\item \href{#method-ReluctantHelperVoice-transform}{\code{ReluctantHelperVoice$transform()}}
\item \href{#method-ReluctantHelperVoice-transform_purpose}{\code{ReluctantHelperVoice$transform_purpose()}}
\item \href{#method-ReluctantHelperVoice-clone}{\code{ReluctantHelperVoice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReluctantHelperVoice-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReluctantHelperVoice-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Reluctant Helper voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReluctantHelperVoice$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReluctantHelperVoice-transform"></a>}}
\if{latex}{\out{\hypertarget{method-ReluctantHelperVoice-transform}{}}}
\subsection{Method \code{transform()}}{
Transform a response using Reluctant Helper voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReluctantHelperVoice$transform(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The original AI response text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReluctantHelperVoice-transform_purpose"></a>}}
\if{latex}{\out{\hypertarget{method-ReluctantHelperVoice-transform_purpose}{}}}
\subsection{Method \code{transform_purpose()}}{
Transform just the purpose content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReluctantHelperVoice$transform_purpose(purpose_lines)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{purpose_lines}}{Character vector with purpose content lines}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed purpose lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReluctantHelperVoice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReluctantHelperVoice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReluctantHelperVoice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
