% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_config}
\alias{save_config}
\title{Save the configuration}
\usage{
save_config(config)
}
\arguments{
\item{config}{The configuration list to save}
}
\value{
Invisibly returns the config
}
\description{
Save the configuration
}
\keyword{internal}
