% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{TimeTravelerVoice}
\alias{TimeTravelerVoice}
\title{Time Traveler Character Voice}
\description{
A character voice that claims to be from the future
}
\keyword{internal}
\section{Super class}{
\code{\link[tldrAI:CharacterVoice]{tldrAI::CharacterVoice}} -> \code{TimeTravelerVoice}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeTravelerVoice-new}{\code{TimeTravelerVoice$new()}}
\item \href{#method-TimeTravelerVoice-transform}{\code{TimeTravelerVoice$transform()}}
\item \href{#method-TimeTravelerVoice-transform_purpose}{\code{TimeTravelerVoice$transform_purpose()}}
\item \href{#method-TimeTravelerVoice-clone}{\code{TimeTravelerVoice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelerVoice-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelerVoice-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Time Traveler voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelerVoice$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelerVoice-transform"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelerVoice-transform}{}}}
\subsection{Method \code{transform()}}{
Transform a response using Time Traveler voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelerVoice$transform(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The original AI response text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelerVoice-transform_purpose"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelerVoice-transform_purpose}{}}}
\subsection{Method \code{transform_purpose()}}{
Transform just the purpose content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelerVoice$transform_purpose(purpose_lines)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{purpose_lines}}{Character vector with purpose content lines}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed purpose lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelerVoice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelerVoice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelerVoice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
