% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{WiseMentorVoice}
\alias{WiseMentorVoice}
\title{Wise Mentor Character Voice}
\description{
A character voice that's thoughtful, patient, and guiding
}
\keyword{internal}
\section{Super class}{
\code{\link[tldrAI:CharacterVoice]{tldrAI::CharacterVoice}} -> \code{WiseMentorVoice}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WiseMentorVoice-new}{\code{WiseMentorVoice$new()}}
\item \href{#method-WiseMentorVoice-transform}{\code{WiseMentorVoice$transform()}}
\item \href{#method-WiseMentorVoice-transform_purpose}{\code{WiseMentorVoice$transform_purpose()}}
\item \href{#method-WiseMentorVoice-clone}{\code{WiseMentorVoice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WiseMentorVoice-new"></a>}}
\if{latex}{\out{\hypertarget{method-WiseMentorVoice-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Wise Mentor voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WiseMentorVoice$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WiseMentorVoice-transform"></a>}}
\if{latex}{\out{\hypertarget{method-WiseMentorVoice-transform}{}}}
\subsection{Method \code{transform()}}{
Transform a response using Wise Mentor voice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WiseMentorVoice$transform(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The original AI response text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WiseMentorVoice-transform_purpose"></a>}}
\if{latex}{\out{\hypertarget{method-WiseMentorVoice-transform_purpose}{}}}
\subsection{Method \code{transform_purpose()}}{
Transform just the purpose content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WiseMentorVoice$transform_purpose(purpose_lines)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{purpose_lines}}{Character vector with purpose content lines}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed purpose lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WiseMentorVoice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WiseMentorVoice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WiseMentorVoice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
